Test bench tests are only run when test-testbench profile is active.
Test bench relies on Selenium and Selenium has limited support for
future and old browsers. This makes it impossible for future users to
build an old version of the project unless they install an old version
of firefox. Something that may be impossible.
To fix this issue, the test bench tests are only run when a user
willingly activates the profile.